Moving option to validator client only
