+++ teku/src/main/java/tech/pegasys/teku/cli/options/ValidatorClientOptions.java
@@ -13,8 +13,6 @@
 
 package tech.pegasys.teku.cli.options;
 
-import static tech.pegasys.teku.validator.api.ValidatorConfig.DEFAULT_VALIDATOR_CLIENT_SSZ_BLOCKS_ENABLED;
-
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
@@ -62,7 +60,8 @@ public class ValidatorClientOptions {
       showDefaultValue = CommandLine.Help.Visibility.ALWAYS,
       arity = "0..1",
       fallbackValue = "true")
-  private boolean validatorClientSszBlocksEnabled = DEFAULT_VALIDATOR_CLIENT_SSZ_BLOCKS_ENABLED;
+  private boolean validatorClientSszBlocksEnabled =
+      ValidatorConfig.DEFAULT_VALIDATOR_CLIENT_SSZ_BLOCKS_ENABLED;
 
   @CommandLine.Option(
       names = {"--Xuse-post-validators-endpoint-enabled"},
@@ -74,6 +73,17 @@ public class ValidatorClientOptions {
       fallbackValue = "true")
   private boolean validatorClientUsePostValidatorsEndpointEnabled =
       ValidatorConfig.DEFAULT_VALIDATOR_CLIENT_USE_POST_VALIDATORS_ENDPOINT_ENABLED;
+
+  @Option(
+      names = {"--Xdvt-integration-enabled"},
+      paramLabel = "<BOOLEAN>",
+      description =
+          "Use DVT endpoints to determine if a distributed validator has aggregation duties.",
+      arity = "0..1",
+      hidden = true,
+      fallbackValue = "true")
+  private boolean dvtSelectionsEndpointEnabled =
+      ValidatorConfig.DEFAULT_DVT_SELECTIONS_ENDPOINT_ENABLED;
 
   public void configure(TekuConfiguration.Builder builder) {
     configureBeaconNodeApiEndpoints();
@@ -87,7 +97,8 @@ public class ValidatorClientOptions {
                     validatorClientUsePostValidatorsEndpointEnabled)
                 .failoversSendSubnetSubscriptionsEnabled(failoversSendSubnetSubscriptionsEnabled)
                 .failoversPublishSignedDutiesEnabled(failoversPublishSignedDutiesEnabled)
-                .sentryNodeConfigurationFile(exclusiveParams.sentryConfigFile));
+                .sentryNodeConfigurationFile(exclusiveParams.sentryConfigFile)
+                .dvtSelectionsEndpointEnabled(dvtSelectionsEndpointEnabled));
   }
 
   private void configureBeaconNodeApiEndpoints() {
