+++ teku/src/main/java/tech/pegasys/teku/cli/options/ValidatorOptions.java
@@ -15,7 +15,6 @@ package tech.pegasys.teku.cli.options;
 
 import static tech.pegasys.teku.networks.Eth2NetworkConfiguration.DEFAULT_VALIDATOR_EXECUTOR_THREADS;
 import static tech.pegasys.teku.validator.api.ValidatorConfig.DEFAULT_DOPPELGANGER_DETECTION_ENABLED;
-import static tech.pegasys.teku.validator.api.ValidatorConfig.DEFAULT_DVT_SELECTIONS_ENDPOINT_ENABLED;
 import static tech.pegasys.teku.validator.api.ValidatorConfig.DEFAULT_SHUTDOWN_WHEN_VALIDATOR_SLASHED_ENABLED;
 import static tech.pegasys.teku.validator.api.ValidatorConfig.DEFAULT_VALIDATOR_IS_LOCAL_SLASHING_PROTECTION_SYNCHRONIZED_ENABLED;
 
@@ -165,16 +164,6 @@ public class ValidatorOptions {
       fallbackValue = "true")
   private boolean shutdownWhenValidatorSlashed = DEFAULT_SHUTDOWN_WHEN_VALIDATOR_SLASHED_ENABLED;
 
-  @Option(
-      names = {"--Xdvt-selections-endpoints-enabled"},
-      paramLabel = "<BOOLEAN>",
-      description =
-          "Use DVT endpoints to determine if a distributed validator has aggregation duties.",
-      arity = "0..1",
-      hidden = true,
-      fallbackValue = "true")
-  private boolean dvtSelectionsEndpointEnabled = DEFAULT_DVT_SELECTIONS_ENDPOINT_ENABLED;
-
   public void configure(TekuConfiguration.Builder builder) {
     builder.validator(
         config ->
@@ -194,8 +183,7 @@ public class ValidatorOptions {
                 .executorThreads(executorThreads)
                 .blockV3enabled(blockV3Enabled)
                 .exitWhenNoValidatorKeysEnabled(exitWhenNoValidatorKeysEnabled)
-                .shutdownWhenValidatorSlashedEnabled(shutdownWhenValidatorSlashed)
-                .dvtSelectionsEndpointEnabled(dvtSelectionsEndpointEnabled));
+                .shutdownWhenValidatorSlashedEnabled(shutdownWhenValidatorSlashed));
     validatorProposerOptions.configure(builder);
     validatorKeysOptions.configure(builder);
   }
